requirements.txt
streamlit
transformers
diffusers
torch
Pillow 